+++ src/main/java/app/community/api/member/controller/MemberController.java
@@ -6,8 +6,8 @@ import app.community.api.member.dto.request.CreateAccountRequest;
 import app.community.api.member.dto.request.LoginRequest;
 import app.community.api.member.dto.request.ModifyInfoRequest;
 import app.community.api.member.dto.response.ShowMemberResponse;
-import app.community.api.member.facade.MemberFacade;
-import app.community.api.member.service.MemberService;
+import app.community.domain.member.facade.MemberFacade;
+import app.community.domain.member.service.MemberService;
 import app.community.global.model.ApiResponse;
 import app.community.global.utils.SessionUtil;
 import lombok.RequiredArgsConstructor;
