+++ src/main/java/app/community/api/member/dto/MemberInfo.java
@@ -1,6 +1,6 @@
 package app.community.api.member.dto;
 
-import app.community.domain.member.Member;
+import app.community.domain.member.entity.Member;
 import app.community.global.enumerated.Role;
 import lombok.Getter;
 import lombok.Setter;
