--- src/main/java/app/community/api/member/repository/MemberRepository.java
@@ -1,8 +1,0 @@
-package app.community.api.member.repository;
-
-import app.community.domain.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Member findByEmailAndPassword(String email, String password);
-}
