--- src/main/java/app/community/api/post/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package app.community.api.post.repository;
-
-import app.community.domain.post.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long>, CategoryRepositoryCustom {
-}
