--- src/main/java/app/community/api/post/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package app.community.api.post.repository;
-
-import app.community.domain.post.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Long>, CommentRepositoryCustom {
-}
