--- src/main/java/app/community/api/post/repository/CommentRepositoryCustom.java
@@ -1,8 +1,0 @@
-package app.community.api.post.repository;
-
-import java.util.List;
-
-public interface CommentRepositoryCustom {
-
-    void deleteInQuery(List<Long> ids);
-}
