--- src/main/java/app/community/api/post/repository/ContentRepository.java
@@ -1,7 +1,0 @@
-package app.community.api.post.repository;
-
-import app.community.domain.post.Content;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ContentRepository extends JpaRepository<Content, Long>, ContentRepositoryCustom {
-}
