--- src/main/java/app/community/api/post/repository/ContentRepositoryCustom.java
@@ -1,8 +1,0 @@
-package app.community.api.post.repository;
-
-import java.util.List;
-
-public interface ContentRepositoryCustom {
-
-    void deleteInQuery(List<Long> ids);
-}
