+++ src/main/java/app/community/domain/member/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package app.community.domain.member.repository;
+
+import app.community.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Member findByEmailAndPassword(String email, String password);
+}
