+++ src/main/java/app/community/domain/post/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package app.community.domain.post.repository;
+
+import app.community.domain.post.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long>, CategoryRepositoryCustom {
+}
