+++ src/main/java/app/community/domain/post/repository/CategoryRepositoryCustom.java
@@ -1,0 +1,10 @@
+package app.community.domain.post.repository;
+
+import app.community.api.post.dto.response.CategoryResponse;
+
+import java.util.List;
+
+public interface CategoryRepositoryCustom {
+
+    List<CategoryResponse> showAll(Long parentId);
+}
