+++ src/main/java/app/community/domain/post/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package app.community.domain.post.repository;
+
+import app.community.domain.post.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long>, CommentRepositoryCustom {
+}
