+++ src/main/java/app/community/domain/post/repository/CommentRepositoryCustom.java
@@ -1,0 +1,8 @@
+package app.community.domain.post.repository;
+
+import java.util.List;
+
+public interface CommentRepositoryCustom {
+
+    void deleteInQuery(List<Long> ids);
+}
