+++ src/main/java/app/community/domain/post/repository/ContentRepository.java
@@ -1,0 +1,7 @@
+package app.community.domain.post.repository;
+
+import app.community.domain.post.entity.Content;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ContentRepository extends JpaRepository<Content, Long>, ContentRepositoryCustom {
+}
