+++ src/main/java/app/community/domain/post/service/CommentService.java
@@ -1,0 +1,36 @@
+package app.community.domain.post.service;
+
+import app.community.domain.post.repository.CommentRepository;
+import app.community.domain.member.entity.Member;
+import app.community.domain.post.entity.Comment;
+import app.community.domain.post.entity.Content;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class CommentService {
+
+    private final CommentRepository commentRepository;
+
+    public Comment findComment(Long id) {
+        return commentRepository.findById(id)
+                .orElseThrow(() -> new NullPointerException("해당 댓글은 존재하지 않는 댓글입니다."));
+    }
+
+    @Transactional
+    public Comment create(String commentContent, Content content, Comment parentComment, Member member) {
+        Comment comment = Comment.createComment(commentContent, content, parentComment, member);
+        return commentRepository.save(comment);
+    }
+
+    @Transactional
+    public void deleteAllComment(List<Long> ids) {
+        if (ids.size() != 0)
+            commentRepository.deleteInQuery(ids);
+    }
+}
