--- src/test/java/app/community/api/post/service/ContentServiceTest.java
@@ -1,27 +1,0 @@
-package app.community.api.post.service;
-
-import app.community.api.member.repository.MemberRepository;
-import app.community.api.post.repository.CategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional
-@SpringBootTest
-class ContentServiceTest {
-
-    @Autowired
-    private ContentService contentService;
-
-    @Autowired
-    private CategoryRepository categoryRepository;
-
-    @Autowired
-    private MemberRepository memberRepository;
-
-
-    private Long categoryId;
-    private Long memberId;
-
-
-}
