+++ src/test/java/app/community/global/test/TestMemberBuilder.java
@@ -1,6 +1,6 @@
 package app.community.global.test;
 
-import app.community.domain.member.Member;
+import app.community.domain.member.entity.Member;
 import app.community.global.enumerated.Role;
 import app.community.global.utils.SHA256Util;
